added an episode generator

+++ csrc/random_video.h
@@ -1,6 +1,8 @@
 #ifndef RANDOM_VIDEO_H
 #define RANDOM_VIDEO_H
 
+#include <stdbool.h>
+
 /* generates a h264 video where 
  * every pixel is sampled from 
  * a discrete uniform distribution 
@@ -10,7 +12,7 @@ bool generate_rv(
     const char *file,
     int width,
     int height,
-    double fps,
+    int fps,
     int gop,
     const char *preset,
     const char *crf,
